feat: support custom relay URL and endpoint
So far the only way to connect a relay was to add a base url (like http://myrelay.com)
with a hardcoded (/admin/manage) endpoint. So to use this app you'd have to download
it and run it locally with your own env variables or deploy it with your env variables
if you wanted to host it.

But if we allow the ability to add/change base url and endpoint and then host this app
somewhere anyone can use it directly without having to build it and then host it or use
it locally.

This opens the possibility for allowing non-technical folks to use it to manage their
relay, and doesn't do anything to restrict from using it locally or hosting it for anyone
who'd like to.