chore: prevent nostr extension from triggering before app is initialized
The extension would ask for permissions for event signinng before the base url
was set. This is due to the fact that the app initialization happened inside
one of the components of the application (Metadata), so the whole component had
to load to render InitializeApp.

This does not make sense as InitializeApp was supposed to take in the input
that's required for initializing the whole application. Loading this component
after part of the app is initialized and inside of another component is wrong.

In this PR, we are improving things by moving this component to a higher level
in the main App component. This way we can also hide the navigation and prevent
rendering of outlets in case someone tries to navigate manually through the browser
navigation bar.