fix(kinds): floating loadData function variable
Because loadData was defined outside of useEffect, but called inside
biome linter correctly pointed out that loadData variable should be
in the dependency array.

I guess this is not the react way because loadData becomes a floating
variable that can change. We can use useCallback and wrap it in and then
put that in the dependency array but we really don't need to here.

Also, removed the loadData calls after updating the state. If the api
call is successful it doesn't make sense to fetch the state again, we
can update it locally