fix(metadata): EditableInput not allowed to be empty / state did not update with the api
Two bugs -
1. If the input was empty (falsy) !inputValue condition would trigger and the original
value would fill in. AKA the input couldn't be empty and rewritten to a new one.

2. EditableInput has a local state to handle the input text, but Metadata passes the
state that's in sync with the API (or we would have to pull the state in EditableInput,
but as I write this maybe that's fine? anyways we send the data from EditableInput?
I will rethink and revisit this.) For now, the function that sends to the API also
updates the state in the parent component