chore(blossom): stop using listallowedpubkeys as a default for fetching pubkeys used to fetch blobs
Added an input to add pubkey and get blobs of that pubkey instead. Using listallowedpubkeys directly
behind the scenes might throw people off, and listallowedpubkeys, seems a bit arbitrary.

This is more explicit and perhaps as (and if) this becomes something that people use we can see
what their preferences are. The ideal UI is perhaps that one that implements multiple ways to
get blobs.

Best way would be to add a BUD to fetch all blobs only if the pubkey for auth event is of the owner's
pubkey of the server.